requirements-test.txt
pre-commit==4.2.0
ruff==0.11.8
pytest==8.3.4
pytest-asyncio==0.23.8
pytest-env==1.1.5
pytest-mock==3.14.0
pytest-cov==6.0.0
sqlalchemy-utils==0.41.2
asyncpg==0.30.0
httpx==0.27.2
